requirements.txt
accelerate==0.31.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
datasets==3.6.0
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.0
exceptiongroup==1.3.0
faiss-gpu==1.7.2
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
gin-config==0.5.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.3
numpy==1.24.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.87.0
packaging==25.0
pandas==1.5.3
pillow==11.0.0
platformdirs==4.3.8
polars==1.9.0
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==3.3.1
sentry-sdk==2.30.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tlparse==0.3.30
tokenizers==0.21.1
torch==2.6.0
torch-geometric==2.6.1
torchaudio==2.5.1+cu121
torchvision==0.21.0
tqdm==4.67.1
transformers==4.52.4
triton==3.2.0
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.4.0
wandb==0.19.0
xxhash==3.5.0
yarl==1.20.1
